🚚 Moves some stuff around
Everything is under a /lib/ folder now, plus
there are now utils etc.
