+++ lib/rules/__tests__/__fixtures__/no-sg-in-components.fixture.js
