+++ lib/rules/__tests__/no-classname-on-common-components.js
