+++ lib/rules/__tests__/no-sg-in-components.spec.js
