+++ lib/rules/no-classname-on-common-components.js
