+++ lib/utils/scope-types.js
@@ -1,0 +1,2 @@
+module.exports.isClass = scope => scope.type === 'class';
+module.exports.isModuleScope = scope => scope.type === 'module';
